+++ client/public/favicon.ico
