+++ client/src/components/list/index.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import TodoItem from '../todo-item';
+import './list.css';
+
+const List = props => {
+
+    const { todos, column, title, editTodo, setEditTodo, updateTodo, deleteTodo } = props;
+
+    return (
+        <div className="list-container">
+            <h3 className="list__title">{title}</h3>
+            {todos ?
+                <ol className="list">
+                    {todos.length === 0 && column === 'created'
+                        ? <p className="placeholder-message">You haven't created any todos yet!</p>
+                        : todos
+                            .filter(item => item.state === column)
+                            .map((item, i) =>
+                                <li key={i}>
+                                    <TodoItem {...item}
+                                        editTodo={editTodo}
+                                        setEditTodo={setEditTodo}
+                                        updateTodo={updateTodo}
+                                        deleteTodo={deleteTodo}
+                                    />
+                                </li>)
+                    }
+                </ol>
+                : <></>}
+        </div>
+    )
+}
+
+export default List;
