+++ client/src/components/welcome/index.js
