+++ client/src/utils/formatTitle.js
